Move LinkTag to MarkupBundle
