+++ src/Sulu/Bundle/MarkupBundle/Markup/Link/LinkConfigurationBuilder.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\PageBundle\Markup\Link;
+namespace Sulu\Bundle\MarkupBundle\Markup\Link;
 
 class LinkConfigurationBuilder
 {
