+++ src/Sulu/Bundle/MarkupBundle/Resources/config/services.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-
     <services>
         <service id="sulu_markup.tag.registry" class="Sulu\Bundle\MarkupBundle\Tag\TagRegistry">
             <argument type="collection"/>
@@ -30,6 +29,16 @@
 
             <tag name="kernel.event_listener" event="kernel.response" method="replaceMarkup" priority="-10"/>
         </service>
-    </services>
 
+        <service id="sulu_markup.link_tag.provider_pool"
+                 class="Sulu\Bundle\MarkupBundle\Markup\Link\LinkProviderPool">
+            <argument type="collection"/>
+        </service>
+
+        <service id="sulu_markup.link_tag" class="Sulu\Bundle\MarkupBundle\Markup\LinkTag">
+            <argument type="service" id="sulu_markup.link_tag.provider_pool"/>
+
+            <tag name="sulu_markup.tag" tag="link" type="html"/>
+        </service>
+    </services>
 </container>
