+++ src/Sulu/Bundle/PageBundle/Resources/config/link-tag.xml
@@ -2,17 +2,6 @@
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <service id="sulu_page.link_tag" class="Sulu\Bundle\PageBundle\Markup\LinkTag">
-            <argument type="service" id="sulu_page.link_tag.provider_pool"/>
-
-            <tag name="sulu_markup.tag" tag="link" type="html"/>
-        </service>
-
-        <service id="sulu_page.link_tag.provider_pool"
-                 class="Sulu\Bundle\PageBundle\Markup\Link\LinkProviderPool">
-            <argument type="collection"/>
-        </service>
-
         <service id="sulu_page.link_tag.page_provider"
                  class="Sulu\Bundle\PageBundle\Markup\Link\PageLinkProvider">
             <argument type="service" id="sulu_page.content_repository"/>
@@ -25,7 +14,7 @@
         </service>
 
         <service id="sulu_page.link_tag.js_config" class="Sulu\Bundle\PageBundle\Admin\LinkProviderJsConfig">
-            <argument type="service" id="sulu_page.link_tag.provider_pool"/>
+            <argument type="service" id="sulu_markup.link_tag.provider_pool"/>
 
             <tag name="sulu.js_config"/>
             <tag name="sulu.context" context="admin"/>
